bots/requirements.txt
amqp==5.0.6
APScheduler==3.6.3
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
binance-connector==1.4.0
bson==0.5.10
cachetools==4.2.2
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==3.4.7
dnspython==2.1.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
joblib==1.0.1
kombu==5.1.0
mongoengine==0.23.1
numpy==1.21.1
pandas==1.3.1
prompt-toolkit==3.0.19
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pymongo==3.12.0
pyOpenSSL==20.0.1
python-dateutil==2.8.2
python-telegram-bot==13.7
pytz==2021.1
redis==3.5.3
requests==2.26.0
scikit-learn==0.24.2
scipy==1.7.1
service-identity==21.1.0
six==1.16.0
sklearn==0.0
ta==0.7.0
threadpoolctl==2.2.0
tornado==6.1
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.0
tzlocal==3.0
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
zope.interface==5.4.0
